recipe: add OpenMP threadprivate test
In bb2a5bd (Fix linking to libgcc_s, 2020-12-21), the libgfortran.spec
file was fixed so that compiling code with 'threadprivate' OpenMP
directives does not require linking with the internal libgcc_eh library.

To prevent future regressions, add a test to the test script
'gfortran_test.sh' that compiles and runs a Fortran program with OpenMP
threadprivate directives. The program is taken from the OpenMP examples
[1].

Closes conda-forge/gfortran_impl_osx-64-feedstock#36

[1] https://github.com/OpenMP/Examples/blob/master/sources/Example_threadprivate.5.f